undo common changes for clutter
