+++ storage/badger/transactions.go
@@ -2,6 +2,7 @@ package badger
 
 import (
 	"github.com/dgraph-io/badger/v2"
+
 	"github.com/onflow/flow-go/model/flow"
 	"github.com/onflow/flow-go/module"
 	"github.com/onflow/flow-go/module/metrics"
